requirements.txt
opencv-python
setuptools
Pillow
numpy